Development

+++ Code/E4_EvaluateRecData.py
@@ -76,9 +76,9 @@ data['MCEvent'] = data['MCEvent'].astype(str)
 data['MCTrack'] = data['MCTrack'].astype(int)
 data['MCTrack'] = data['MCTrack'].astype(str)
 data['MC_Track_ID'] = data['MCEvent'] + '-' + data['MCTrack']
-data['quarter'] = data['quarter'].astype(int)
+#data['quarter'] = data['quarter'].astype(int)
 data['quarter'] = data['quarter'].astype(str)
-data['FEDRATrackID'] = data['FEDRATrackID'].astype(int)
+#data['FEDRATrackID'] = data['FEDRATrackID'].astype(int)
 data['FEDRATrackID'] = data['FEDRATrackID'].astype(str)
 data['FEDRA_Track_ID'] = data['quarter'] + '-' + data['FEDRATrackID']
 data=data.drop(['MCEvent','MCTrack','VertexS','PID','MotherID','FEDRATrackID','quarter'],axis=1,inplace=True)
